// src/app/api/admin/article-categories/article-category.type.ts
import { Document, Model, Types } from 'mongoose';

export type TArticleCategoryStatus = 'active' | 'inactive';

export type TArticleCategorySEO = {
  title?: string;
  description?: string;
  keywords?: string[];
};

export type TArticleCategory = {
  sequence: number;
  icon?: string;
  thumbnail?: string;
  name: string;
  slug: string;
  description?: string;
  status: TArticleCategoryStatus;
  tags: string[];
  parent?: Types.ObjectId | null;
  layout?: string;
  seo?: TArticleCategorySEO;
  is_deleted?: boolean;
};

export interface TArticleCategoryDocument
  extends TArticleCategory,
    Document {
  _id: Types.ObjectId;
  softDelete(): Promise<TArticleCategoryDocument | null>;
}

export type TArticleCategoryModel = Model<TArticleCategoryDocument> & {
  isCategoryExist(_id: string): Promise<TArticleCategoryDocument | null>;
};